Warning refactors
[Refactor]: Discard unused anonymous method arguments
[Refactor]: Remove unused using directives
[Refactor]: Remove unnecessary null deniers (!)
[Modification]: Add usage of TextureCache in TextureHelper

+++ Onlooker/Common/2D/Vector2Int.cs
@@ -65,7 +65,7 @@ public record struct Vector2Int(int X, int Y)
     
     public static implicit operator Point(Vector2Int vector)
     {
-        return new Point((int)vector.X, (int)vector.Y);
+        return new Point(vector.X, vector.Y);
     }
     
     public override string ToString()
